now the Z-matrix is adjusted in the 'aggregate.py' file; the labeler returns the matrices X, T and Z; changed some minor parts

+++ knodle/labeler/CheXpert/stages/z_matrix.py
@@ -16,6 +16,6 @@ def z_matrix():
                           names=[REPORTS])[REPORTS].tolist()
     n_samples = len(reports)
 
-    Z_matrix = np.zeros((n_samples, n_rules), dtype=int)
+    Z_matrix = np.zeros((n_samples, n_rules))  # , dtype=int
 
     return Z_matrix
